==> pipeline/3.1-profile-data.ipynb <==
import geopandas as gpd

tract_file = "../data/raw/tracts.geojson"

tracts = gpd.read_file(tract_file)

# Cardinality

assert not tracts.geoid10.duplicated().any()
assert not tracts.geometry.duplicated().any()
assert not tracts['centroid'].duplicated().any()

print("We expect {} tracts".format(tracts.geoid10.nunique()))
# Chicago itself has 801 tracts.
