==> pipeline/5.3-reduce-divvy.ipynb <==
import geopandas as gpd

from data.geo import wkb_geom
from data.util import coalesce

bike_rides_in = "../data/interim/bike_rides.geoparquet"

bike_rides_out = "../data/interim/bike_rides_v2.geoparquet"

# Pipeline In

bike_rides = gpd.read_parquet(bike_rides_in).pipe(wkb_geom, 'geometry_imputed')

# Simplify schema

Use the imputed id name and geom, which have lower cardinality.

bike_rides = bike_rides.pipe(coalesce, 'station_id_imputed','station_id', 'station_id')
bike_rides = bike_rides.pipe(coalesce, 'station_name_imputed','station_name', 'station_name')
bike_rides = bike_rides.pipe(coalesce, 'geometry_imputed','geometry', 'geometry')

To reduce risk of future confusion, I'll drop the station_id and vintage and geometry columns.
The station name is useful in identifying clusters and useful descriptively so we keep it,
even though it is not a PK.

bike_rides = bike_rides.drop(columns=['station_id','vintage'])

bike_rides = (bike_rides
              .groupby(['geometry','station_name','date'], as_index=False)
              .agg({"start_rides": lambda x: x.fillna(0).sum(),
                    "end_rides": lambda x: x.fillna(0).sum(),
                    "rides": lambda x: x.fillna(0).sum()})
              .pipe(gpd.GeoDataFrame, geometry='geometry'))

assert all(bike_rides.groupby('station_name')['geometry'].nunique() == 1 )

# there's only a couple geoms that point to different station names but it's just mis-spellings. 
bike_rides[bike_rides.groupby('geometry')['station_name'].transform('nunique') > 1][['geometry',	'station_name']].drop_duplicates()

# Pipeline Out

bike_rides.to_parquet(bike_rides_out, index=False)
