==> pipeline/5.4-profile-data.ipynb <==
import geopandas as gpd
import pandas as pd
from plotly import express as px

from data.datemath import date_range_daily

train_rides_in = "../data/raw/train_rides.csv"
bus_rides_in = "../data/raw/bus_rides.csv"
bike_rides_in = "../data/interim/bike_rides_v2.geoparquet"
uber_rides_in = "../data/raw/uber_tract_rides.parquet"

train_file_in = "../data/raw/train_stations.geojson"
bus_routes_file_in = "../data/raw/bus_routes.geojson"
bike_stations_file_in = "../data/raw/bike_stations.geojson"
tract_file_in = "../data/raw/tracts.geojson"

train_rides = pd.read_csv(train_rides_in)
bus_rides = pd.read_csv(bus_rides_in)
bike_rides = gpd.read_parquet(bike_rides_in)
uber_rides = pd.read_parquet(uber_rides_in)

train_stations = gpd.read_file(train_file_in)
bus_routes = gpd.read_file(bus_routes_file_in)
tract_points = gpd.read_file(tract_file_in)
bike_stations = gpd.read_file(bike_stations_file_in)

# Cardinality

bike_rides[['station_name','geometry']].nunique()

# Ride Units not in Stations Units

### Train

assert train_rides.station_id.isin(train_stations.map_id).all()

### Uber

excess = ~uber_rides.tract.isin(tract_points.geoid10)
print("{} uber tracts aren't in chicago".format(
    uber_rides[excess]['tract'].nunique()
))
# We already know the uber dataset contains some rides TO or FROM chicago.

### Bus

excess = ~bus_rides.route.isin(bus_routes.route)
print("These bus rides aren't in the routes list: ",
      set(bus_rides[excess]['route']))

### Bike

We know the trip data isn't in the station data because the later vintages are denormalized.

# Units not in Rides

### Train

assert train_stations.map_id.isin(train_rides.station_id).all()

### Uber

print("Uber data has {} / {} cook county tracts".format(
    tract_points.geoid10.isin(uber_rides.tract).sum(),
    tract_points.geoid10.nunique()
))

### Bus

print("Are all bus routes present in ridership? ", 
      bus_routes.route.isin(bus_rides.route).all())

missing = bus_routes[~bus_routes.route.isin(bus_rides.route)]
missing
# We are missing one route. That's fine. 
# We'll implicitly drop it when we merge on ridership.
# This is unfortunate because that route runs near mccormick place.
# It only runs during bears games though which happened on Aug 1, 10, 17, 22

### Bike

print("Latest stations vintage: ",
      bike_stations['vintage'].str.extract(r"(\d{4})").max().iloc[0])

# Listen, the relevant ridership years are all denormalized
# ==> we already have station locations in the ridership tables directly.
# We don't even want this data because we only want post-pandemic data at the most.
# ==> don't even need to merge these stations to ridership.
# ==> we won't even check this query which takes a few sec to run anyway
# assert bike_stations.geometry.isin(bike_rides.geometry).all()

# Dates

## All dates

dates = set(map(lambda x: x[0], date_range_daily("2024-07-01","2024-08-31")))

assert all([d in train_rides['date'].values for d in dates])
assert all([d in bus_rides['date'].values for d in dates])
assert all([d in uber_rides['date'].values for d in dates])
assert all([d in bike_rides['date'].values for d in dates])

## Units x Dates

### Train

for g, gdf in train_rides.groupby('station_id'):
    assert all([d in gdf['date'].values for d in dates])

plot_data = train_rides.groupby('date')['station_id'].nunique().rename('completeness') / train_rides['station_id'].nunique()
px.line(plot_data.reset_index(), x='date', y='completeness')
# The green line damen station comes online during the study period, so we will need
# to impute its prior ridership.

### Bus

compliance = bus_rides.groupby('route')['date'].agg(lambda x: sum([d in x.values for d in dates]))
compliance = compliance / len(dates)
compliance.describe()
# We have a 50% panel of ridership for bus rides

plot_data = bus_rides.groupby('date')['route'].nunique().rename('completeness') / bus_rides['route'].nunique()
px.line(plot_data.reset_index(), x='date', y='completeness')
# Three things mess with the panel:
#  - some routes operate every day and some only on weekdays
#  - a small number of routes are added/removed from service over time
#  - the panel never records 0 ridership days so those migth be implied

### Uber

compliance = uber_rides.groupby('tract')['date'].agg(lambda x: sum([d in x.values for d in dates]))
compliance = compliance / len(dates)
compliance.describe()
# We have a 50% panel of ridership for uber tracts


plot_data = uber_rides.groupby('date')['tract'].nunique().rename('completeness') / uber_rides['tract'].nunique()
px.line(plot_data.reset_index(), x='date', y='completeness')
# I think two things mess with the panel:
#   - we know that ridership is much higher on weekends than weekdays
#   - people go to certain tracts dichotomously on weekends vs weekdays
#   - the panel doesn't display 0 ridership days so these might be implied
#   - the panel anonymizes low ridership days to community area level so 
#     as far as we are ignoring those we are treating them as 0 ridership too.
# The mean completeness probably saturates over time due to market forces.

### Bike

compliance = bike_rides.groupby(['station_name'])['date'].agg(lambda x: len(set(x) & dates))
compliance = compliance / len(dates)
compliance.describe()
# We have a 75% panel of ridership for bike stations

plot_data = bike_rides.groupby('date')['station_name'].nunique().rename('completeness')
plot_data = plot_data / bike_rides['station_name'].nunique()
px.line(plot_data.reset_index(), x='date', y='completeness')
