==> pipeline/6.1-spatial-features.ipynb <==
import geopandas as gpd
import pandas as pd
import numpy as np
from shapely.geometry import Point

from data.constants import (WORLD_CRS, LOCAL_CRS,
                                OHARE_CENTROID, MIDWAY_CENTROID)
from data.geo import meter_to_foot, dms_to_decimal, foot_to_meter

train_file_in = "../data/raw/train_stations.geojson"
bus_stops_file_in = "../data/raw/bus_stops.geojson"
tract_file_in = "../data/raw/tracts.geojson"
comm_file_in = "../data/raw/communities.geojson"
poi_file_in = "../data/raw/poi_buildings.geojson"

# The merge between bike stations and bike rides was really complicated (5.x notebooks)
# Now we operate on the bike_rides data because it has correct imputed geoms.
bike_rides_in = "../data/interim/bike_rides_v2.geoparquet"

train_stations_out = "../data/interim/train_stations.geojson"
train_lines_out = "../data/interim/train_lines.csv"
bus_stops_file_out = "../data/interim/bus_stops.geojson"
tract_file_out = "../data/interim/tracts.geoparquet"
comm_file_out = "../data/interim/communities.geojson"

bike_rides_out = "../data/interim/bike_rides_v3.geoparquet"

TOLERANCE = 200  # meters

# Pipeline in

train_stations = gpd.read_file(train_file_in)
bus_stops = gpd.read_file(bus_stops_file_in)
tract_points = gpd.read_file(tract_file_in)
comm_points = gpd.read_file(comm_file_in)
poi_buildings = gpd.read_file(poi_file_in)

bike_rides = gpd.read_parquet(bike_rides_in)

# Time-Invariant Features

## Airports

### -> CTA

We can label up the train and bus stops via name.

train_stations['airport'] = (train_stations.station_name == "O'Hare") \
                            | (train_stations.station_name == "Midway")
train_stations['airport'] = train_stations['airport'].astype(float)

# Note: We only check for Midway because CTA busses don't go directly into O'Hare,
#       nor even to the adjacent Mixed Modal Transit center.
bus_stops['airport'] = (bus_stops['PUBLIC_NAME'] == "Midway Orange Line Station").astype(float)

### -> Uber

The rideshare pickups are anonymized to census area so we can't use building catchements.

oh_xy = dms_to_decimal(*OHARE_CENTROID[1]), dms_to_decimal(*OHARE_CENTROID[0])
oh_xy = Point(*oh_xy) # lng/lat
oh_xy = gpd.GeoSeries([oh_xy], crs=WORLD_CRS).to_crs(LOCAL_CRS).iloc[0]
oh_tract = tract_points.set_index('geoid10').geometry.to_crs(LOCAL_CRS).contains(oh_xy)
oh_comm = comm_points.set_index('comm_area').geometry.to_crs(LOCAL_CRS).contains(oh_xy)

mdw_xy = dms_to_decimal(*MIDWAY_CENTROID[1]), dms_to_decimal(*MIDWAY_CENTROID[0])
mdw_xy = Point(*mdw_xy) # lng/lat
mdw_xy = gpd.GeoSeries([mdw_xy], crs=WORLD_CRS).to_crs(LOCAL_CRS).iloc[0]
mdw_tract = tract_points.set_index('geoid10').geometry.to_crs(LOCAL_CRS).contains(mdw_xy)
mdw_comm = comm_points.set_index('comm_area').geometry.to_crs(LOCAL_CRS).contains(mdw_xy)

tract_points['airport'] = (tract_points['geoid10'].map(oh_tract) | tract_points['geoid10'].map(mdw_tract)).astype(float)
comm_points['airport'] = (comm_points['comm_area'].map(oh_comm) | comm_points['comm_area'].map(mdw_comm)).astype(float)

### -> Divvy

We're going to assume that no one rides a bike to the airport for out of town travel.

bike_rides['airport'] = 0.0

## United Center, McCormick Place catchements

uc_building = poi_buildings[poi_buildings['name'] == 'UNITED CENTER']
mp_building = poi_buildings[poi_buildings['name'] == 'HYATT REGENCY MCCORMICK PLACE']

def buffered_intersects(gdf: gpd.GeoDataFrame, geom:gpd.GeoSeries, geom_prefix:str, dists:list[int]):
    """
    Computes multiple buffer distances around geom and EQUIVALENTLY computes:
        - whether each row of gdf intersects buffer
        - whether buffer intersects each row of gdf
    Params:
        - gdf: compares each row of gdf to geom
        - geom: the shape to compute buffers around. EXPECTS SINGLE ROW
        - geom_prefix: name for resulting buffer comparison column
        - dists: list of buffer radii in meters
    """
    assert len(geom) == 1, "Expects only one geom to buffer around."
    building_proj = geom.geometry.to_crs(LOCAL_CRS)
    buffers = [building_proj.buffer(meter_to_foot(d)) for d in dists]
    df_proj = gdf.geometry.to_crs(LOCAL_CRS)
    codes = {f"{geom_prefix}_{d}": df_proj.intersects(b.iloc[0]) * 1.0 for d,b in zip(dists,buffers)}
    return gdf.assign(**codes)

**Coded as** *"is this station within the catchement of the POI"*

# Stations are coded as 
train_stations = train_stations.pipe(buffered_intersects, uc_building, "uc", [400,800,1600])
bike_rides = bike_rides.pipe(buffered_intersects, uc_building, "uc", [400,800,1600])
bus_stops = bus_stops.pipe(buffered_intersects, uc_building, "uc", [400,800,1600])

train_stations = train_stations.pipe(buffered_intersects, mp_building, "mp", [400,800,1600])
bike_rides = bike_rides.pipe(buffered_intersects, mp_building, "mp", [400,800,1600])
bus_stops = bus_stops.pipe(buffered_intersects, mp_building, "mp", [400,800,1600])

**Coded as** *"is this POI catchement intersecting the tract"*

tract_points = tract_points.pipe(buffered_intersects, uc_building, "uc", [400,800,1600])
comm_points = comm_points.pipe(buffered_intersects, uc_building, "uc", [400,800,1600])

tract_points = tract_points.pipe(buffered_intersects, mp_building, "mp", [400,800,1600])
comm_points = comm_points.pipe(buffered_intersects, mp_building, "mp", [400,800,1600])

## Tract Membership

Label the station tracts so we can aggregate to tract-level to match uber data.

def code_tract(gdf):
    """
    Spatial join point to census tract
    """
    tracts_proj = tract_points.to_crs(LOCAL_CRS)
    gdf_proj = gdf.to_crs(LOCAL_CRS)
    comm_proj = comm_points.to_crs(LOCAL_CRS)
    coded_tracts = gdf_proj[['geometry']].sjoin(tracts_proj[['geoid10','geometry']], how='left', predicate='within')
    coded_comms = gdf_proj[['geometry']].sjoin(comm_proj[['comm_area','geometry']], how='left', predicate='within')
    gdf = pd.concat([gdf, 
                    coded_tracts['geoid10'].rename('tract'), 
                    coded_comms['comm_area'].rename('comm_area')],
                    axis=1)
    return gdf

train_stations = train_stations.pipe(code_tract)
bike_rides = bike_rides.pipe(code_tract)
bus_stops = bus_stops.pipe(code_tract)

print("Bike rides with unknown tract {:.2%}".format(
    bike_rides.tract.isna().mean()
))

At this point 0.2% of bike rides are not strictly within a tract. 
But they happen to be really close, so let's

def code_nearest_tract(gdf):
    """
    Spatial join point to census tract
    """
    tracts_proj = tract_points.to_crs(LOCAL_CRS)
    gdf_proj = gdf.to_crs(LOCAL_CRS)
    comm_proj = comm_points.to_crs(LOCAL_CRS)
    opts = dict(how='left', distance_col='dist', max_distance=meter_to_foot(TOLERANCE))
    coded_tracts = gdf_proj[['geometry']].sjoin_nearest(tracts_proj[['geoid10','geometry']], **opts)
    coded_comms = gdf_proj[['geometry']].sjoin_nearest(comm_proj[['comm_area','geometry']], **opts)
    assert all(coded_tracts.dist <= TOLERANCE)
    assert all(coded_comms.dist <= TOLERANCE)
    gdf = pd.concat([gdf, 
                    coded_tracts['geoid10'].rename('tract'), 
                    coded_comms['comm_area']],
                    axis=1)
    return gdf

bike_rides_nearest = bike_rides[bike_rides.tract.isna()].drop(columns=['tract','comm_area'])
bike_rides_nearest = code_nearest_tract(bike_rides_nearest)
# Just take first in case any are equidistant.
bike_rides_nearest = bike_rides_nearest.loc[bike_rides_nearest.index.drop_duplicates()]
assert bike_rides_nearest[['tract','comm_area']].notna().all().all()

bike_rides = pd.concat([bike_rides[bike_rides.tract.notna()], 
                        bike_rides_nearest])

## Transit access by tract

def access_stats(area_df, transit_df, transit_name, area_id, transit_id):
    area_centroids = (area_df
                      .assign(centroid = area_df.geometry.to_crs(LOCAL_CRS).centroid)
                      .set_geometry('centroid')
                      .drop(columns=['geometry'])
                      .filter([area_id,'centroid']))
   
    # Note sjoin_nearest is m:m if points are equidistant ==> we further consolidate
    access = area_centroids.sjoin_nearest(transit_df[['geometry']].to_crs(LOCAL_CRS), 
                                           how='left', 
                                           distance_col=f'{transit_name}_distance') \
                        .groupby(area_id,as_index=False)[f'{transit_name}_distance'].min()

    contained = area_df.merge(transit_df, left_on=area_id, right_on=transit_id, how='left') \
                        .groupby(area_id,as_index=False).size() \
                        .rename(columns={'size':f'{transit_name}_contained'})
    
    area_df = area_df.merge(access).merge(contained)
    return area_df

tract_points = tract_points.pipe(access_stats, train_stations[['tract','geometry']], 'train', 'geoid10', 'tract')
tract_points = tract_points.pipe(access_stats, bike_rides[['tract','geometry']], 'bike', 'geoid10', 'tract')
tract_points = tract_points.pipe(access_stats, bus_stops[['tract','geometry']], 'bus', 'geoid10', 'tract')

comm_points = comm_points.pipe(access_stats, train_stations[['comm_area','geometry']], 'train', 'comm_area', 'comm_area')
comm_points = comm_points.pipe(access_stats, bike_rides[['comm_area','geometry']], 'bike', 'comm_area', 'comm_area')
comm_points = comm_points.pipe(access_stats, bus_stops[['comm_area','geometry']], 'bus', 'comm_area', 'comm_area')

# Yes it's silly putting self-distance, but it just makes merging the frames easier later.
train_stations = train_stations.pipe(access_stats, train_stations[['tract','geometry']], 'train', 'tract', 'tract')
train_stations = train_stations.pipe(access_stats, bike_rides[['tract','geometry']], 'bike', 'tract', 'tract')
train_stations = train_stations.pipe(access_stats, bus_stops[['tract','geometry']], 'bus', 'tract', 'tract')
bike_rides = bike_rides.pipe(access_stats, train_stations[['tract','geometry']], 'train', 'tract', 'tract')
bike_rides = bike_rides.pipe(access_stats, bike_rides[['tract','geometry']], 'bike', 'tract', 'tract')
bike_rides = bike_rides.pipe(access_stats, bus_stops[['tract','geometry']], 'bus', 'tract', 'tract')

## Voronoi Area

voronoi = train_stations.to_crs(LOCAL_CRS).voronoi_polygons().clip(tract_points.to_crs(LOCAL_CRS).envelope)
train_stations = train_stations.assign(landarea = voronoi.area.apply(np.sqrt).apply(foot_to_meter))

tract_points = tract_points.assign(landarea = tract_points.to_crs(LOCAL_CRS).area.apply(np.sqrt).apply(foot_to_meter))
comm_points = comm_points.assign(landarea = comm_points.to_crs(LOCAL_CRS).area.apply(np.sqrt).apply(foot_to_meter))

bike_ride_stations = bike_rides.geometry.drop_duplicates()
voronoi = bike_ride_stations.to_crs(LOCAL_CRS).voronoi_polygons().clip(tract_points.to_crs(LOCAL_CRS).envelope)
voronoi = dict(zip(bike_ride_stations, voronoi))
bike_rides = bike_rides.assign(landarea = bike_rides.geometry.map(voronoi).area.apply(np.sqrt).apply(foot_to_meter))

# Pipeline out

train_stations.to_file(train_stations_out, index=False)
bus_stops.to_file(bus_stops_file_out, index=False)
tract_points.to_parquet(tract_file_out, index=False)
comm_points.to_file(comm_file_out, index=False)

bike_rides.to_parquet(bike_rides_out, index=False)
