==> pipeline/7.1-profile-data.ipynb <==
import pandas as pd

point_panel_in = "../data/interim/point_panel.parquet"
line_panel_in = "../data/interim/line_panel.parquet"
tract_panel_in = "../data/interim/tract_panel.parquet"

# Pipeline in

point_panel = pd.read_parquet(point_panel_in)
line_panel = pd.read_parquet(line_panel_in)
tract_panel = pd.read_parquet(tract_panel_in)

# Cardinality

## Units

pd.concat([point_panel.groupby('transit')['id'].nunique().rename('point'),
            line_panel.groupby('transit')['id'].nunique().rename('line'),
            tract_panel.groupby('transit')['id'].nunique().rename('tract')],
            axis=1)
# This is ok.

## Date

pd.concat([point_panel.groupby('transit')['date'].nunique().rename('point'),
            line_panel.groupby('transit')['date'].nunique().rename('line'),
            tract_panel.groupby('transit')['date'].nunique().rename('tract')],
            axis=1)
# This is ok too.

# Nulls

# TODO: DEBUG:
# Where are these nulls coming from?
point_panel.groupby('transit')[['tract','lat','long']].agg(lambda x: x.isna().mean())
