==> pipeline/8.1-model-data.ipynb <==
import pandas as pd
import numpy as np

point_panel_in = "../data/final/point_panel.parquet"
tract_panel_in = "../data/final/tract_panel.parquet"

panel_out = "../data/final/model_data.parquet"

# Pipeline In

point_panel = pd.read_parquet(point_panel_in)
tract_panel = pd.read_parquet(tract_panel_in)

tract_panel = tract_panel.loc[~tract_panel.is_weekend]
point_panel = point_panel.loc[~point_panel.is_weekend]

def get_model_data(panel, transit):
    model_data = panel.query(f'transit == "{transit}"')
    empty_units = model_data.groupby('id')['rides'].transform('sum') == 0
    model_data = model_data[~empty_units]
    return model_data

model_data = pd.concat([
    get_model_data(tract_panel, "uber").assign(id = lambda x: x['id'].astype(str)),
    get_model_data(point_panel, "train").assign(id = lambda x: x['id'].astype(str)),
    get_model_data(point_panel, "bike").assign(id = lambda x: x['id'].astype(str)),
])

# Add log distance.
for col in model_data.columns:
    if 'distance' in col:
        model_data['log_' + col] = model_data[col].apply(np.log1p)

# Pipeline Out

model_data.to_parquet(panel_out, index=False)
