==> pipeline/10.1-profile-data.ipynb <==
import geopandas as gpd
import os
import pandas as pd
from plotly import express as px

from data.constants import (DATA_FOLDER)
from data.datemath import date_range_daily

point_panel_in = os.path.join(DATA_FOLDER, "final", "point_panel.parquet")
line_panel_in = os.path.join(DATA_FOLDER, "final", "line_panel.parquet")
tract_panel_in = os.path.join(DATA_FOLDER, "final", "tract_panel.parquet")

# Pipeline in

point_panel = pd.read_parquet(point_panel_in)
line_panel = pd.read_parquet(line_panel_in)
tract_panel = pd.read_parquet(tract_panel_in)

# Validate

TODO! XXX!

Add assertions here about data completeness, like:

- do we have a full time panel for every unit?
- do we have the expected number of stations, lines, tracts
- do we have the expected number of transits per tract
- are our datatypes valid? are binary variables actually binary (i actually made this mistake)
